Create README.md
Papua New Guinea (PNG) is known for its breathtaking but challenging landscape, characterized by rugged terrain and vast mountainous regions with multiple creeks and rivers crossing everywhere. Developing infrastructure in such an environment is no small feat. Adding to this challenge, climate change continues to accelerate infrastructure damage and deterioration, resulting in millions of Kina in losses annually. This not only undermines development efforts but also leads to infrastructure that lacks longevity and sustainability.
To address these pressing issues, the integration of Geospatial Technology and engineering-based knowledge is crucial for climate-resilient infrastructure planning, design, and construction. Essential infrastructure such as road networks, bridges, and energy systems plays a critical role in improving living standards and driving economic growth. Ensuring their resilience against environmental challenges is imperative for the country's sustainable development.
Across both urban and rural regions of PNG, millions of Kina are often spent on infrastructure projects that, unfortunately, fail to withstand the test of time, deteriorating within months or just a few years. One of the key reasons behind this issue is the lack of proper or adequate baseline data to guide construction. Without a strong data-driven foundation, infrastructure projects are prone to short lifespans, increased maintenance costs, and vulnerability to climate-related impacts.
Over the past five to six years, my research team and I, with support from the School of Surveying and Land Studies and the Faculty of Postgraduate Studies, Research & Innovation at PNG University of Technology, have been actively engaged in developing sustainable and climate-resilient infrastructure development solutions. Our research has primarily focused on the Mape Catchment in the Finschhafen District, serving as a model site for generating innovative solutions that can be adapted and applied to other regions of county PNG.